installer: CPDUP removes -vvv -I
While here, reduce the override overhead.  bsdinstaller is in its
final resting place now and ideally we only need a single config
file.  The only trouble is that core.git will have to adapt to the
config files present to load the correct ones without having the
App throw errors on startup.